new import API: bulk, scan
