+++ src/main/scala/hammerlab/iterators/package.scala
@@ -1,0 +1,5 @@
+package hammerlab
+
+package object iterators
+  extends bulk
+    with scan
