+++ src/main/scala/org/hammerlab/iterator/GroupWithIterator.scala
@@ -1,6 +1,6 @@
 package org.hammerlab.iterator
 
-import org.hammerlab.iterator.bulk.BufferedBulkIterator._
+import hammerlab.iterators.bulk._
 
 /**
  * Group one sorted iterator with another, emitting an iterator of the latter's elements for each of the former's
